New Java Files, Movement & Rotation, and more...
I wanted to make UML diagrams for my code so that I could get the structure situated ahead of time. I did do this, but not really in UML. Kind of just wrote it out in a garbled and messy way that probably only makes sense to me, but whatever works!

-

I created some Java files based on the code I sketched out:

Asset (Class)
will store object-specific data (position, size, location, etc.).

InteractableAsset (Abstract Class) (extends Asset)
For objects that CAN interact with other objects, BE interacted with BY other objects, or BOTH. I included three abstract methods in here that, when defined, should work together to process interactions. My goal for this is to have the InteractableAsset player object be able to interact with the InteractableAsset bench object by pressing 'e', or by clicking, or some other means of interaction. Then, because they are both InteractableAssets, the sending/receiving of the interaction is handled by the code defined in the three methods.

- Updatable (Interface)
 This interface simply includes three method declarations; load, update, and render. Assets or InteractableAssets should implement this interface if they want to have a manipulatable state. I plan for each of these methods to be called for every object that implements this interface in the corresponding libGDX create and render methods. Without implementing this interface, an Asset or InteractableAsset may remain static (no movement, no animation, no position updates, etc.)

Player (Class) (extends InteractableAsset) (implements Updatable)
This will hold all of the relevant player information (eatingSpeed, eatingRepuation, position, size, etc.). The player is an InteractableAsset because the user will walk up to booths to initiate training or a speed record attempt. The player implement updatable because it will not be static. There will be animations, movement, and other state updates.

TrainingBooth (Class) (extends InteractableAsset)
This will hold all of the relevant training booth information. The training booth is a static InteractableAsset because it won't move or have animations, but it will be interacted with by the player.

ChallengeBooth (Class) (extends InteractableAsset)
This will hold all the relevant information for training booths. These will be structured similarly to training booths, but I think they will have enough differences to the point where they need their own classes. If that doesn't end up being the case, I might make it into one class where I use a boolean or something to differentiate whether it is a training or challenge booth.

Food (Class) (extends Asset) (implements Updatable):
This will be an instantiable class that holds all food related data. It extends Asset because it won't need to interact with any other assets or be interacted with. I plan for these to be used as props primarily, so they won't typically have their state altered. But I do hope to implement animations for the food so that it looks like bites are being taken out of them, and if I end up being able to do that, then they need to implement the Updatable methods so that their animation can be altered.

Animation (Class)
This will handle the animations for each object. Not sure if I'm going to actually make this class or if I'm going to use a class or library from libGDX, but I added it now just in case.

-

I worked on movement and rotation of the starter donut asset we have, and its coming along nicely. I experimented after following a tutorial video and I think I know how to properly fix the camera into a first-person angle. Or, I could figure out how to modify the FirstPersonCameraController so that it isn't able to move on the y-axis, because that already handles the fixating of the camera and then I'd only have x and z axis movement.

-

Renamed the project from KitchenLife to Worlds Best Speed Eater Simulator (because thats how they name them on Roblox). Also organized and removed some files.