Foundation Laid & File System Changes
I went through all of the starter code and pulled out the parts that I'll need for my own project. This was a long but very helpful process, as I now have a better understanding of the libGDX classes that I'm working off of.

I made the Asset class so that it is much easier to get Blender models into the game. Also reorganized some file system stuff.