Lots of Code Updates, Working on the Logic
Working on the logic for the actual gameplay, mainly between the Booth and Player class. It's going well, and I think I have a good event processing circuit going from start to end with the restaurant record run and training runs. Only way to initiate them will be through the interaction from Player to Booth.

I am going to have Ryan make Key model assets, as in keyboard keys. I plan to have them float up behind the food as its being eaten at random or on a timer. The player can press the key before it disappears to get some "bonus bites" in. The bonus bites will be the total number of keys pressed in a second multiplied by 10% of the Player's eating speed.

Although this game is simple, I'm still having a lot of fun making it and I'm enjoying the creation of the gameplay logic code.