Training Booth and Challenge Booth Logic
I did a bunch of stuff to various areas of code. Not an ounce of work that I just did had any sense of order or chronology to it. Whenever something popped into my head, I just went after it, so there's probably some unfinished stuff in there. But I think it's going well, just gotta deal with the collision detection next which I have been procrastinating for sure.