Text Pop-Ups In-Game, Booths, Commenting, and more
I got the Drawer class a bit more functional. It now displays text on the screen.

I also made it so that you can sit in the Booth, look around without your position changing, and stand up smoothly. I think I might make it so that once you sit, you can freely look around UNTIL you press 'e' again to start eating. Idk though.

All code files are perfectly organized and commented. Trying to maintain that well-documented state as I progress, but sometimes I fall behind and have to spend an hour or so organizing and commenting. It is very effective though, and I know that because I just relied on my Killer code to help me remember how to get 2D drawings and fonts and stuff working and my documentation back then was very neat and organized so I was able to do that easily.