Coordinate System Work
Ryan and I spent a good hour and a half figuring out how the coordinate system from Blender translates to libGDX. We figured out that when we export things from Blender, we need to export them from (0, 0, 0), right at the origin.

Each model in Blender will have a different location, lets say (5, 5, 5). When it's exported from Blender at (5, 5, 5), when I load it into the game at (0, 0, 0) it will spawn at (5, 5, 5). I was unaware of that being the case, but we figured it out and adjusted to it.