Arms Rotation Attempts
EULER ANGLES:
Euler angles are a method used to rotate objects into new orientations. Euler angles take a pitch (usually X), yaw (usually Y), and roll (usually Z) rotation value (in degrees or radians) and will use those values to rotate an object into its desired new orientation.

Objects rotate on the plane formed by the other two axis
Rotate On X-Axis: rotates on YZ plane
Rotate On Y-Axis: rotates on XZ plane
Rotate On Z-Axis: rotates on XY plane

Orientation: word describing where the object is pointing at
Rotation: act of altering an Object's orientation

Gimbal Lock: when an adjustment to one axis can cause the other two axis to be on the same plane (and therefore modifying one of those two is the same as modifying the other). This can be bad and cause some loss of precision in rotations.

QUATERNIONS:
Quaternions provide a solution to a problem that euler angles encounter. When given the same inputs for rotation values (degrees or radians to rotate), using euler angles could result in various different orientations after rotation. That's because the order of axis rotation matters when using euler angles.

With quaternions, regardless of the order of axis rotation, when given the same input values for rotation they will will always produce the same orientation.

Quaternions are represented by four values, typically X, Y, Z, W.
X, Y, and Z in this instance are not representing the axises, they are representing their own individual values like W.

Quaternions give weight to each of the 4 values (between 0 to 1), and they will rotate according to that in a four dimensional space.